// src/generated/SessionReliefParams.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ReliefAction } from "./ReliefAction";
import type { SessionId } from "./SessionId";

/**
 * `session.relief` params.
 */
export type SessionReliefParams = { session_id: SessionId, action: ReliefAction, };